requirements.txt
# Core dependencies
flask>=3.0
flask-limiter>=3.5
apiflask>=2.0
pydantic>=2.0
pillow>=10.0
python-magic>=0.4
gunicorn>=23.0
python-dotenv>=1.0
google-genai>=1.0
google-api-core>=2.0
langgraph>=0.2
langchain-google-genai>=2.0
langchain-core>=0.3
pyjwt>=2.8
requests>=2.31

# Web tools
httpx>=0.27
beautifulsoup4>=4.12
html2text>=2024.2
ddgs>=7.0

# Code execution sandbox
llm-sandbox[docker]>=0.1

# Language detection
langdetect>=1.0

# Schedule parsing
croniter>=5.0

# Database migrations
yoyo-migrations>=9.0

# Development dependencies
ruff>=0.8
mypy>=1.13
types-requests>=2.31
types-croniter>=5.0

# Testing
pytest>=8.0
pytest-cov>=4.0
pytest-mock>=3.0
openapi-core>=0.19
